+++ pkg/directory/azure/config.go
@@ -1,0 +1,69 @@
+package azure
+
+import (
+	"net/http"
+	"net/url"
+)
+
+const (
+	defaultGraphHost = "graph.microsoft.com"
+
+	defaultLoginHost      = "login.microsoftonline.com"
+	defaultLoginScope     = "https://graph.microsoft.com/.default"
+	defaultLoginGrantType = "client_credentials"
+)
+
+type config struct {
+	graphURL       *url.URL
+	httpClient     *http.Client
+	loginURL       *url.URL
+	serviceAccount *ServiceAccount
+}
+
+// An Option updates the provider configuration.
+type Option func(*config)
+
+// WithGraphURL sets the graph URL for the configuration.
+func WithGraphURL(graphURL *url.URL) Option {
+	return func(cfg *config) {
+		cfg.graphURL = graphURL
+	}
+}
+
+// WithLoginURL sets the login URL for the configuration.
+func WithLoginURL(loginURL *url.URL) Option {
+	return func(cfg *config) {
+		cfg.loginURL = loginURL
+	}
+}
+
+// WithHTTPClient sets the http client to use for requests to the Azure APIs.
+func WithHTTPClient(httpClient *http.Client) Option {
+	return func(cfg *config) {
+		cfg.httpClient = httpClient
+	}
+}
+
+// WithServiceAccount sets the service account to use to access Azure.
+func WithServiceAccount(serviceAccount *ServiceAccount) Option {
+	return func(cfg *config) {
+		cfg.serviceAccount = serviceAccount
+	}
+}
+
+func getConfig(options ...Option) *config {
+	cfg := new(config)
+	WithGraphURL(&url.URL{
+		Scheme: "https",
+		Host:   defaultGraphHost,
+	})(cfg)
+	WithHTTPClient(http.DefaultClient)(cfg)
+	WithLoginURL(&url.URL{
+		Scheme: "https",
+		Host:   defaultLoginHost,
+	})(cfg)
+	for _, option := range options {
+		option(cfg)
+	}
+	return cfg
+}
